@@ -1,13 +1,15 @@
-
-import { Pipe, PipeTransform } from '@angular/core';
+import { Pipe, PipeTransform, inject } from '@angular/core';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 
 @Pipe({
   name: 'safeUrl'
 })
-export class SafeUrlPipe implements PipeTransform {
-  constructor(private sanitizer: DomSanitizer) {}
-  transform(url: string): SafeResourceUrl {
+export class SafeUrlPipe implements PipeTransform 
+{
+  private sanitizer = inject(DomSanitizer);
+  
+  transform(url: string): SafeResourceUrl 
+  {
     return this.sanitizer.bypassSecurityTrustResourceUrl(url);
   }
 }
