@@ -1,12 +1,16 @@
 import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-root',
   template: `
     <router-outlet></router-outlet>
   `,
-  styles: []
+  styles: [],
+  standalone: true,
+  imports: [RouterOutlet]
 })
-export class AppComponent {
+export class AppComponent 
+{
   title = 'myhelp';
 }
